fixed a bug in how seqNo was being handled between subscribe calls.

+++ noojeecontact.api/src/main/java/au/org/noojee/contact/api/NoojeeContactApi.java
@@ -411,7 +411,7 @@ public class NoojeeContactApi
 		@SerializedName("Data")
 		HashMap<String, List<Event>> endPointEventMap;
 
-		int seq;
+		long seq;
 		int Code;
 
 		public List<EndPointEvent> getEvents()
